Only use conf.maldet.cron for overriding cron settings.
I intended on overriding any maldet configuration settings with
conf.maldet.cron. It turns out that file is only sourced in the daily cron job.

If you want to override maldet configuration settings that aren't related to
the cron task, you need to do it in conf.maldet. It is the only configuration
file that is sourced when starting the maldet service.

I am closing pull request 20, now that I know how this works.
https://github.com/nexcess/puppet-maldet/pull/20